refractored the routes
