+++ Express/AuthDemo/views/home.ejs
@@ -7,4 +7,4 @@
         
 </div>
 
-<% include partials/footer %>
+<% include partials/footer %>
